E-commerce
In traditional system a client  has  move to market  in order to get her shopping . There is no guarantee that the client will shop his/her  that day. .There is no doubt that in this system a lot of  shopping bag  are needed to be handled. Markets  have to various shop and  people .Client bargaining the price .Some times client foolish the shop man . Along with this the client  sometimes lose their money bag and his/her energy . This age old system is also very much time consuming and eventually lessen the quality of service.  The system will have different level of user. The one who is not registered as a user will only be able to view the basic information in the side but can not shopping in the side.